pyproject.toml
[tool.poetry]
name = "webtronics"
version = "0.1.0"
description = ""
authors = ["samsegomof <wemagonof91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
poetry = "^1.5.1"
fastapi = "^0.99.1"
sqlalchemy = "^2.0.18"
alembic = "^1.11.1"
asyncpg = "^0.27.0"
python-dotenv = "^1.0.0"
greenlet = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

